[Fix] Community 관련 API를 v2로 업데이트

+++ app/src/main/java/com/kust/kustaurant/domain/usecase/community/DeleteCommunityPostUseCase.kt
@@ -8,7 +8,7 @@ import javax.inject.Singleton
 class DeleteCommunityPostUseCase @Inject constructor(
     private val communityRepository: CommunityRepository
 ){
-    suspend operator fun invoke(postId : Int) {
+    suspend operator fun invoke(postId : Long) {
         communityRepository.deletePost(postId)
     }
 }
